requirements.txt
aiogram==3.4.1
yt-dlp==2024.04.09
python-dotenv==1.0.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
qrcode==7.4.2
Pillow==10.4.0
aihorde==2.0.2